requirements.txt
websockets==15.0.1
fastapi
uvicorn
PyJWT
webauthn
pydantic
